+++ sdk/core/azure-core-http-netty/src/main/java/com/azure/core/http/netty/NettyAsyncHttpClientBuilder.java
@@ -16,12 +16,13 @@ import com.azure.core.util.Configuration;
 import com.azure.core.util.Context;
 import com.azure.core.util.CoreUtils;
 import com.azure.core.util.logging.ClientLogger;
+import io.netty.buffer.PooledByteBufAllocator;
 import io.netty.channel.ChannelOption;
 import io.netty.channel.EventLoopGroup;
-import io.netty.channel.nio.NioEventLoopGroup;
 import io.netty.handler.logging.LoggingHandler;
 import io.netty.resolver.DefaultAddressResolverGroup;
 import io.netty.resolver.NoopAddressResolverGroup;
+import io.netty.util.internal.SystemPropertyUtil;
 import reactor.netty.Connection;
 import reactor.netty.NettyPipeline;
 import reactor.netty.http.HttpDecoderSpec;
@@ -151,7 +152,7 @@ public class NettyAsyncHttpClientBuilder {
      *     .wiretap&#40;TcpClient.class.getName&#40;&#41;, LogLevel.INFO&#41;;
      * &#47;&#47; Create an HttpClient based on above reactor-netty client and configure EventLoop count.
      * HttpClient client = new NettyAsyncHttpClientBuilder&#40;baseHttpClient&#41;
-     *     .eventLoopGroup&#40;new NioEventLoopGroup&#40;5&#41;&#41;
+     *     .eventLoopGroup&#40;new MultiThreadIoEventLoopGroup&#40;5, NioIoHandler.newFactory&#40;&#41;&#41;&#41;
      *     .build&#40;&#41;;
      * </pre>
      * <!-- end com.azure.core.http.netty.from-existing-http-client -->
@@ -200,6 +201,12 @@ public class NettyAsyncHttpClientBuilder {
             nettyHttpClient = HttpClient.create(this.connectionProvider);
         } else {
             nettyHttpClient = HttpClient.create();
+        }
+
+        // Keep the pooled default while respecting allocator choices made by the application.
+        if (!nettyHttpClient.configuration().options().containsKey(ChannelOption.ALLOCATOR)
+            && !SystemPropertyUtil.contains("io.netty.allocator.type")) {
+            nettyHttpClient = nettyHttpClient.option(ChannelOption.ALLOCATOR, PooledByteBufAllocator.DEFAULT);
         }
 
         // If a resolver hasn't been set, set the default one.
@@ -385,12 +392,12 @@ public class NettyAsyncHttpClientBuilder {
     /**
      * Sets the NIO event loop group that will be used to run IO loops.
      *
-     * @param nioEventLoopGroup The {@link NioEventLoopGroup} that will run IO loops.
+     * @param nioEventLoopGroup The {@link io.netty.channel.nio.NioEventLoopGroup} that will run IO loops.
      * @return the updated NettyAsyncHttpClientBuilder object.
      * @deprecated deprecated in favor of {@link #eventLoopGroup(EventLoopGroup)}.
      */
     @Deprecated
-    public NettyAsyncHttpClientBuilder nioEventLoopGroup(NioEventLoopGroup nioEventLoopGroup) {
+    public NettyAsyncHttpClientBuilder nioEventLoopGroup(io.netty.channel.nio.NioEventLoopGroup nioEventLoopGroup) {
         this.eventLoopGroup = nioEventLoopGroup;
         return this;
     }
@@ -404,7 +411,7 @@ public class NettyAsyncHttpClientBuilder {
      * <pre>
      * int threadCount = 5;
      * HttpClient client = new NettyAsyncHttpClientBuilder&#40;&#41;
-     *     .eventLoopGroup&#40;new NioEventLoopGroup&#40;threadCount&#41;&#41;
+     *     .eventLoopGroup&#40;new MultiThreadIoEventLoopGroup&#40;threadCount, NioIoHandler.newFactory&#40;&#41;&#41;&#41;
      *     .build&#40;&#41;;
      * </pre>
      * <!-- end com.azure.core.http.netty.NettyAsyncHttpClientBuilder#eventLoopGroup -->