+++ sdk/core/azure-core-http-netty/src/samples/java/com/azure/core/http/netty/NettyAsyncHttpClientBuilderJavaDocCodeSnippets.java
@@ -7,7 +7,8 @@ import com.azure.core.http.HttpClient;
 import com.azure.core.http.HttpMethod;
 import com.azure.core.http.HttpRequest;
 import com.azure.core.http.ProxyOptions;
-import io.netty.channel.nio.NioEventLoopGroup;
+import io.netty.channel.MultiThreadIoEventLoopGroup;
+import io.netty.channel.nio.NioIoHandler;
 import io.netty.handler.logging.LogLevel;
 import reactor.netty.resources.ConnectionProvider;
 import reactor.netty.tcp.TcpClient;
@@ -39,7 +40,7 @@ public class NettyAsyncHttpClientBuilderJavaDocCodeSnippets {
         // BEGIN: com.azure.core.http.netty.NettyAsyncHttpClientBuilder#eventLoopGroup
         int threadCount = 5;
         HttpClient client = new NettyAsyncHttpClientBuilder()
-            .eventLoopGroup(new NioEventLoopGroup(threadCount))
+            .eventLoopGroup(new MultiThreadIoEventLoopGroup(threadCount, NioIoHandler.newFactory()))
             .build();
         // END: com.azure.core.http.netty.NettyAsyncHttpClientBuilder#eventLoopGroup
     }
@@ -87,7 +88,7 @@ public class NettyAsyncHttpClientBuilderJavaDocCodeSnippets {
             .wiretap(TcpClient.class.getName(), LogLevel.INFO);
         // Create an HttpClient based on above reactor-netty client and configure EventLoop count.
         HttpClient client = new NettyAsyncHttpClientBuilder(baseHttpClient)
-            .eventLoopGroup(new NioEventLoopGroup(5))
+            .eventLoopGroup(new MultiThreadIoEventLoopGroup(5, NioIoHandler.newFactory()))
             .build();
         // END: com.azure.core.http.netty.from-existing-http-client
     }