+++ sdk/formrecognizer/azure-ai-formrecognizer/src/test/java/com/azure/ai/formrecognizer/documentanalysis/administration/DocumentModelAdminClientTest.java
@@ -29,7 +29,6 @@ import com.azure.core.test.http.AssertingHttpClientBuilder;
 import com.azure.core.util.BinaryData;
 import com.azure.core.util.Context;
 import com.azure.core.util.polling.SyncPoller;
-import io.netty.handler.codec.http.HttpResponseStatus;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -110,7 +109,7 @@ public class DocumentModelAdminClientTest extends DocumentModelAdministrationCli
                 = client.getDocumentModelWithResponse(documentModelDetails.getModelId(), Context.NONE);
             client.deleteDocumentModel(documentModelDetails.getModelId());
 
-            assertEquals(documentModelResponse.getStatusCode(), HttpResponseStatus.OK.code());
+            assertEquals(200, documentModelResponse.getStatusCode());
             validateDocumentModelData(documentModelResponse.getValue());
         });
     }
@@ -134,7 +133,7 @@ public class DocumentModelAdminClientTest extends DocumentModelAdministrationCli
         DocumentAnalysisServiceVersion serviceVersion) {
         client = getDocumentModelAdministrationClient(httpClient, serviceVersion);
         Response<ResourceDetails> resourceDetailsResponse = client.getResourceDetailsWithResponse(Context.NONE);
-        assertEquals(resourceDetailsResponse.getStatusCode(), HttpResponseStatus.OK.code());
+        assertEquals(200, resourceDetailsResponse.getStatusCode());
         validateResourceInfo(resourceDetailsResponse.getValue());
     }
 
@@ -166,7 +165,7 @@ public class DocumentModelAdminClientTest extends DocumentModelAdministrationCli
             final Response<Void> deleteModelWithResponse
                 = client.deleteDocumentModelWithResponse(createdModel.getModelId(), Context.NONE);
 
-            assertEquals(deleteModelWithResponse.getStatusCode(), HttpResponseStatus.NO_CONTENT.code());
+            assertEquals(204, deleteModelWithResponse.getStatusCode());
             final HttpResponseException exception = assertThrows(HttpResponseException.class,
                 () -> client.getDocumentModelWithResponse(createdModel.getModelId(), Context.NONE));
             final ResponseError responseError = (ResponseError) exception.getValue();