+++ sdk/servicebus/azure-messaging-servicebus/src/test/java/com/azure/messaging/servicebus/administration/ServiceBusAdministrationAsyncClientTest.java
@@ -38,7 +38,6 @@ import com.azure.messaging.servicebus.administration.models.QueueProperties;
 import com.azure.messaging.servicebus.administration.models.QueueRuntimeProperties;
 import com.azure.xml.XmlSerializable;
 import com.azure.xml.XmlWriter;
-import io.netty.handler.codec.http.HttpResponseStatus;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -528,7 +527,7 @@ class ServiceBusAdministrationAsyncClientTest {
 
         // Act & Assert
         StepVerifier.create(client.deleteRuleWithResponse(topicName, subscriptionName, ruleName))
-            .assertNext(actual -> assertEquals(actual.getStatusCode(), HttpResponseStatus.NO_CONTENT.code()))
+            .assertNext(actual -> assertEquals(204, actual.getStatusCode()))
             .expectComplete()
             .verify(DEFAULT_TIMEOUT);
     }
@@ -555,7 +554,7 @@ class ServiceBusAdministrationAsyncClientTest {
 
         // Act & Assert
         StepVerifier.create(client.deleteSubscriptionWithResponse(topicName, subscriptionName))
-            .assertNext(actual -> assertEquals(actual.getStatusCode(), HttpResponseStatus.NO_CONTENT.code()))
+            .assertNext(actual -> assertEquals(204, actual.getStatusCode()))
             .expectComplete()
             .verify(DEFAULT_TIMEOUT);
     }
@@ -581,7 +580,7 @@ class ServiceBusAdministrationAsyncClientTest {
 
         // Act & Assert
         StepVerifier.create(client.deleteTopicWithResponse(topicName))
-            .assertNext(actual -> assertEquals(actual.getStatusCode(), HttpResponseStatus.NO_CONTENT.code()))
+            .assertNext(actual -> assertEquals(204, actual.getStatusCode()))
             .expectComplete()
             .verify(DEFAULT_TIMEOUT);
     }