+++ sdk/webpubsub/azure-messaging-webpubsub-client/src/main/java/module-info.java
@@ -9,6 +9,7 @@ module com.azure.messaging.webpubsub.client {
     requires io.netty.transport;
     requires io.netty.handler;
     requires io.netty.codec;
+    requires io.netty.codec.compression;
     requires io.netty.codec.http;
 
     exports com.azure.messaging.webpubsub.client;